final service commit

+++ internal/models/subscription.go
@@ -3,9 +3,9 @@ package models
 import "time"
 
 type Subscription struct {
-	ID          int64      `json:"id" db:"id"`
+	ID          int        `json:"id" db:"id"`
 	ServiceName string     `json:"service_name" db:"service_name"`
-	Price       int64      `json:"price" db:"price"`
+	Price       int        `json:"price" db:"price"`
 	UserID      string     `json:"user_id" db:"user_id"`
 	StartDate   time.Time  `json:"start_date" db:"start_date"`
 	EndDate     *time.Time `json:"end_date" db:"end_date"`
@@ -13,7 +13,7 @@ type Subscription struct {
 
 type CreateSubscriptionRequest struct {
 	ServiceName string  `json:"service_name"`
-	Price       int64   `json:"price"`
+	Price       int     `json:"price"`
 	UserID      string  `json:"user_id"`
 	StartDate   string  `json:"start_date"`
 	EndDate     *string `json:"end_date"`
@@ -21,7 +21,7 @@ type CreateSubscriptionRequest struct {
 
 type UpdateSubscriptionRequest struct {
 	ServiceName string  `json:"service_name"`
-	Price       int64   `json:"price"`
+	Price       int     `json:"price"`
 	StartDate   string  `json:"start_date"`
 	EndDate     *string `json:"end_date"`
 }
